+++ lib/src/eihidef.h
@@ -1,6 +1,8 @@
 #ifndef EIHIDEF_H
 #define EIHIDEF_H
 
+#include <vms_types.h>
+
 #define EIHI$K_MAJORID 1
 #define EIHI$K_MINORID 2
 #define EIHI$K_LENGTH 104
@@ -11,10 +13,10 @@ struct _eihi
 {
     struct
     {
-        unsigned int eihi$l_majorid;
-        unsigned int eihi$l_minorid;
+        UINT32 eihi$l_majorid;
+        UINT32 eihi$l_minorid;
     };
-    unsigned long long eihi$q_linktime;
+    UINT64 eihi$q_linktime;
     char eihi$t_imgnam [40];
     char eihi$t_imgid [16];
     char eihi$t_linkid [16];
