+++ lib/src/ihidef.h
@@ -1,6 +1,8 @@
 #ifndef IHIDEF_H
 #define IHIDEF_H
 
+#include <vms_types.h>
+
 #define IHI$K_LENGTH 80
 #define IHI$C_LENGTH 80
 #define IHI$S_IHIDEF 80
@@ -9,7 +11,7 @@ struct _ihi
 {
     char ihi$t_imgnam [40];
     char ihi$t_imgid [16];
-    unsigned long long ihi$q_linktime;
+    UINT64 ihi$q_linktime;
     char ihi$t_linkid [16];
 };
